added warning, updated permissions tests, added permissions

+++ server/app/api.py
@@ -762,6 +762,10 @@ class AssignmentAPI(APIResource):
             raise BadValueError(err)
         
     def delete(self, obj, user, data):
+        need = Need('delete')
+        if not obj.can(user, need, obj):
+            raise need.exception()
+
         obj.key.delete()
